requirements.txt
requests==2.31.0
python-dotenv==1.0.0
pymongo==4.6.1
loguru>=0.7.2
pyyaml>=6.0.1
brotli>=1.1.0 